Support -t and -d options to findup, when python not available. I.E. make fixdup.sh script support report only and delete modes.
git-svn-id: https://fslint.googlecode.com/svn/trunk@64 494acc53-a219-0410-adff-8f8ff343f446

+++ fslint/supprt/rmlint/fixdup.sh
@@ -4,15 +4,47 @@
 
 suffix=.fixdup.$$
 
+link=1
+dryRun=0
+if [ "$#" = "1" ]; then
+    if [ "$1" = "del" ]; then
+        link=0
+    elif [ "$1" = "tdel" ]; then
+        link=0
+        dryRun=1
+    elif [ "$1" = "tmerge" ]; then
+        dryRun=1
+    fi
+fi
+
 keepfile=''
 while read file; do
   if [ -z "$file" ]; then
     keepfile="nextfile"
   elif [ "nextfile" = "$keepfile" ]; then
     keepfile="$file"
+    if [ "$dryRun" = "1" ]; then
+        echo -ne "\nkeeping $keepfile\t"
+        if [ "$link" = "1" ]; then
+            echo -ne "hardlinking: "
+        else
+            echo -ne "deleting: "
+        fi
+    fi
   else
-    ln -f -b --suffix="$suffix" -- "$keepfile" "$file" 2>/dev/null ||
-    ln -sf -b --suffix="$suffix" -- "$keepfile" "$file"
-    rm -f "$file$suffix"
+    if [ "$dryRun" = "1" ]; then
+        echo -ne "$file "
+    else
+        if [ "$link" = "1" ]; then
+            ln -f -b --suffix="$suffix" -- "$keepfile" "$file" 2>/dev/null ||
+            ln -sf -b --suffix="$suffix" -- "$keepfile" "$file"
+            rm -f "$file$suffix"
+        else
+            rm -f "$file"
+        fi
+    fi
   fi
 done
+if [ "$dryRun" = "1"  -a "$keepfile" ]; then
+    echo
+fi
